Fixed EditBackground failing to show "save as" for shortcuts.
Fixed the case where selecting a shortcut does not show the "save as" dropdown, as the fileType is set to application/x-ms-shortcut Not "video".
Thus would completely hide the save as function, meaning you could only ever save it as itself, unless you manually re-selected the target file the shortcut resolved to.