Fixed save as: "copy file" for shortcut paths.
Fixed the problem where a selected shortcut would only try to save itself instead of the resolved path when the "Save Background as" was set to "Copy file"

Now when trying to save with the save as set to copy file it will properly resolve the shortcut, and copy the target file. You can still use save as shortcut and it will create a new shortcut to the original file.