Save Background as... now defaults to Saved type (or Copy file).
Now onMount loads the filetype of the background file. On drag/drop or edit. Note: actual background file must be a video type.

For instance if you drop a shortcut file to create a new background it will default to save shortcut. If you drop the raw video it will default to copy file. Further if you edit, and already have the background saved as a shortcut it defaults to shortcut, and if it is saved as a copy file it defaults to copy file.

Also changed the default in general to Copy File. Yes shortcut is nice because it saves storage space as it does not create a copy of the file, but it also is a much more safe way to store the file in general. I.e. if the original video gets moved/deleted it will still work as it has it locally, unlike with the shortcuts which work fine, till the original file is deleted/moved.

So I have moved Shortcut from being recommended to instead setting Copy file as the default.