External Paths import as new/saved date rework.
Not 100% done with this, so be careful changing external paths. To be safe either add or remove one thing MAX at a time, do not just edit one path.

Added ipc call "indexBackgrounds(options?: { newExternalPathAdded?: boolean })
Called via:
await window.electron.indexBackgrounds(); (to just re-index backgrounds.
OR:
await window.electron.indexBackgrounds({newExternalPathAdded: true,});
Which tells the indexBackgrounds to check for new external paths.
This flag is not so much that it only looks for external paths to add to index, but rather its a flag that tells it "hey i just updated my external paths setting so there might be a new external path to check" Then if there are any bg.json files found in this index it prompts the user to import backgrounds as brand new(show on page 1) or import from their last saved index.
This allows the users to effectively cut/paste a bunch of files from their base directory, or an external directory to a new directory, then add it as an external path. Upon which it gives them the option to import these backgrounds as new, or import them as saved date (allows them to move the base directory but not mess up the order of all of their backgrounds).

Further changes:
All other indexBackgrounds no longer show the "Import as new/Import with saved date". They all default to import as new.

I'm not 100% set in stone on this implementation, for example this is one problem.
Changing external file paths do not correctly re-index unless you only add/remove 1 at a time, changing "E:/test/" to "E:/test2/" does not correctly re-index as expected, any same name IDs do not re-index and instead keep their "E:/test/" date. I think this might be solved just by controlling how the External Paths are saved in general.