Detect background file moves from/to external paths
Added a section in indexBackgrounds that now logs suspected moves from 1 file location to another.
Basically it looks for backgrounds which have the same base ID but were removed from 1 file path and added in another file path.
Effectively this is tracking if the user moves the background from 1 directory to another directory.

The end goal of this is to copy over the indexed time such that it does not re-index this moved background as a brand new background, thus messing up the users ordering.