Transfer indexed date value to suspected move.
Now when a file is moved from 1 directory to another and indexed, it will get the old value (indexed time which controls where it appears in the background select), and restore this value to the new id.
This ensures that even if the user directly moves the files themselves it will maintain it's order.

The only real way to mess this up is to remove it from a directory to a non-external path folder, then open background select/ launch program. Then place it into a directory and upon all of that it will index as new.
So keep the files within external paths/default directory before indexing.
You can move them to non-external path directory then add that path as an external path and it will give you the option to import as new or import with saved date (which will maintain the order).

TL;DR: Move files directly from 1 to another before opening the program/Background Select. Or a new directory then add the directory as an external path and select "import with saved date".