Ensures unique id on moveToBackgroundFolder
Ensures that the moved background obtains a unique ID, on moving it.
Previously if you moved say id: "000001" to External path 0, which had a folder named "000001" it would not move the file and would return null. Now it attempts to name it "000001_1" "000001_2" etc.

Known issue:
When moving it and it gets this new ID, it does not correctly detect it as a suspected move as detecting suspected moves requires the same ID. Will look into fixing this.