Fix _number renames not recognized as suspected moves.
Now when moving and it renames the file to id_1, id_2 etc. it will still detect this as a possible move if "id" was deleted, but "id_1" was added for example.

This is a little bit sloppy and has 1 direct way to cause a problem. Basically it is going way out of ones way to make it work, but it is possible to open BackgroundSelect, then delete a file say "name" then move for instance "name_1" to another directory, and it detects a suspected move with not name_1 but "name" itself, and updates "name_1" to the indexed time of "name". This is highly highly highly unlikely but possible. I might eventually just swap over from copying the indexed time directly to just assuming that the moves will always be import as saved date.