Fix suspected moves desyncing backgrounds.json + bg.json local.indexed.
Fixed a scenario when a background was moved, upon indexing as a new background (before detecting the move) it updates it's local.indexed to the current time.
Now when it detects the background as a suspected move, it reverts the local.indexed time to the same time as it now stores in backgrounds.json.

This better ensures that both the backgrounds.json and the local.indexed are on the same page to avoid problems where importing does import with the local.indexed time. (such as adding new external path and choosing import with saved date).