Delay local.indexed updating till after suspected moves found.
Now delays the local.index in bg.jsons for newly created backgrounds on indexing until after it has been decided if it is a suspected move or not.
This saves us some writes, then re-writes. Also sets up for the plan I have which is to instead of getting the old indexed time for the suspected move from the backgrounds.json to just get it from the bg.json instead, which would be impossible with the old setup because it would have already been overwritten.