Fixed base name causing problems with indexed.
Workflow for setting indexed time for suspected moves has been improved.

It now checks it's full id(file name) for a match from removed, instead of just looking directly for the base name. This helps when you move multiple files like say "name", "name_1", "name_2". Previously this would end up giving the same time for "name_1" and "name_2" since they both matched the base name of the first removed "name" background.

There is still 1 way to mess this up, and it is so stupid that I would have to believe someone did it on purpose, it is the special case of special cases. (besides just copying a file and renaming it to something different, this is always expected to be indexed as new).