Fixed externalPaths not indexing from scratch.
Fixed a problem where externalPaths would no longer index fresh, and only index an extenral path directory if it was directly modified in settings "external paths".

With this all known issues besides extreme extreme special cases which would be too compute heavy to specifically solve for the 1 in 1 trillion chance a user actually does it seem to be fixed.
In 99.99% of all cases it should be working as expected even with manual moving (not ideal, but attempted to support as much as I can).
Now just need to add the UI moving from default directory to external/ in between externals etc. (Already implemented the backend for it just missing the UI options).