Fixed getExternalPath drawing from out of date Backgrounds.json
Fixed a case where I realized it is possible to improperly get backgrounds.json external paths.

I think the future move is to just not store externalPaths in backgrounds.json as it's just a mess and can be desynced from setting.json.

This desync meant that getExternalPath would not find the externalPaths[index] thus would return null which would then fallback to try and just append the ext::0::name as a default directory file. Which would at least not get written, but should not have been in a case where it attempted to write to it regardless.