Small css + Added public tag "Other" (for resolution).
Changed the color for Edit Background category headers to match the color of the edit background panel. (easier to see).

Also added an "Other" for resolution in public tags. Seems to be working fine, not 100% sure this is all it takes for adding a public tag but seems to work fine.