Added names.json + tags.json files (indexes). (Fully for testing)
Now creates a names.json and a tags.json in the base folder.

Mostly just looking to try to ensure that the writing to/from the .jsons are optimized for large background libraries.

For instance my 5800 background test setup already takes 12 seconds to index from scratch. (re-indexing is instant). But this only has 5800 backgrounds which are copied from a background with no name, and no tags.

I am first going to do some rough testing to see if any bottlenecks appear or are fixed etc. Going to create a new test set with tags and names to also test.