Fix accidental page nulling in BackgroundSelect.
Not sure why this popped up again, but popped up with a new background test set (4300 backgrounds with names, and multiple tags).

Regardless made it such that fetching page -1 is no longer possible, it just returns.

I feel like I already fixed multiple areas where it could call with page = -1 on load to instead not call fetchPage(), but for some reason its doing it again with a new massive set...