EditBackground now utilizes the "show" for Local Tags:
Local Tags now utilizes the "show" value on opening to expand/collapse the Local Tags section.

Also updates the "show" on changes to remember the last time you opened this window and keep the same expanded/collapsed state.

Known issue: Category named "show" now appears in EditCategories and EditBackground, when it shouldn't (At least the special case of trying to add a category named "show" fails as it thinks its already a category). Also toggling this does not immediately close the Local tags but does changed the save state (should not appear at all so shouldn't be an issue).