removed "show" categegory from user UI.
"show" category is used for LocalTag categories, and controls whether or not the entire localTags section is expanded/collapsed. It is not a category to be shown to the user so it has been removed.

Still some special cases to handle to ensure users don't create a tag "show"/"Show" etc.