Update publicCategories setting for expanded/collapsed in Filter.
The Filter panel now saves expanded/collapsed states on expandeding/collapsing categories.
For instance collapsing category "Resolution" means that when you open EditBackground it will be in the same state (Resolution is collapsed).

This means that the public categories are now synced between both the Filter Panel and the EditBackground window.
Note: The categories section itself though is not synced, as it does not save yet, only the categories within Public Tags at the moment. Rest will be updated soon to be in sync.