Fixed "" (no category) not appearing in FilterPanel.
Fixed the filter accidentally not displaying No category category due to it not being stored as a direct category.

Last second pre-commit note:
Realized this only happens when the "" category itself is missing, which in normal use it is not missing. I.e. collapsing/expanding it adds it as a category so it should always show. I'm not sure what I did to end up deleting the "" category from the settings, will run some tests to try and find where it was accidentally removed from settings.