Tag suggestions now work for "-tag:"
Suggestions now work for search bar typing "-tag:", the same as they work for typing "tag:". It will select tags which match the partial after the "-tag:" and you can hit Tab, Enter, or click the suggestions to complete the tag query.

"-tag:" is supported in general, but does not currently have a way to apply it via the filter search panel. Might add one eventually. But this "-tag:Video" for example will always be a possibility.