IndexBackgrounds after it loads from saved backgrounds.
BackgroundSelect instead of instantly calling indexBackgrounds and delaying the launch of the window now launches with the saved backgrounds, and then indexes and refreshes the pages after initial loading.

Previously if it took to long to index it would not show the new backgrounds, and would require changing search, or pages in order to show the newly added/removed backgrounds.