Large Refactor + Migration (categories -> localCategories).
Originally did not plan on saving a publicCategories. But ended up adding it.
Now it seems weird to have "categories" referring to just localCategories. So it has been changed to "localCategories"

Added a migration to convert users "categories" to "localCategories" automatically.

So a name change, and for the user nothing needs to be done just launch the program and it will auto migrate to the new name.