updated openBackgroundSelect ipc to take optional id.
Updated it to take an optional id: string parameter, this will be used as an override for scrollTo on opening.
For reference the scrollTo is when BackgroundSelect window is opened it auto scrolls to the selected background, this would instead be a way to make it scroll to a non-selected background.

This would be ideal for when you save but don't apply the background, so you end up back where you were when you edited it, but its not forcing you into setting any background you edit as the current background.