Closing EditBackground now resumes location in BackgroundSelect.
Now when closing EditBackground from editing one, it will re-open BackgroundSelect with the background the user edited set as the scroll to target (goes to the page and scrolls to the background the user just edited). This works like resuming where the user left off instead of force scrolling back to their current set background if they saved, but did not apply, or closed the EditBackground window.