Further improved window size on smaller monitors. (backgroundSelect)
Now ensures that the BackgroundSelectWindow is smaller than the mainWindow. (unless the mainWindow is smaller than 645x450 or the monitor is smaller than 645x450).

This better ensures that the window that pops up doesn't completely hide the unMaximized window and lead to confusion.

Has a fallback to screen dimensions in case the mainWindow dimensions fail for some reason.