Fix dropdowns down arrow not reserving space.
Fixed edit-bg-field overriding the .dropdown-container select{}
and not reserving space for the down arrow on long select options.

Also noticed that the clearable inputs are not 100% as expected, they also have this for all ones not within a <div className="search-menu"> for the time being. Will have to work out how to fix that...