Added "Dim Background" slider to Display -> All Icon Boxes.
Now added a slider "Dim Background" to the legend which controls how dimmed the background becomes. Allowing the user to swap between a fully black background and a fully transparent background so they can more easily see the colors, regardless of the colors of their actual background item. This defaults to 50% on showing.