Legend toggles now work (All Icon Boxes).
Clicking to toggle the different legend fields now works.

All default to being shown, but you can now click to toggle their group from showing. For example if you only want to see the ones with an offset you can disable the oversized and the default and it will only show you items that are offset, and offset + oversized.