Fixed devEnv/ electron build.
Ensured both tsconfigs now use ES2020.
Updated npm run "build" to use proper electron backend compilation. Previously was using tsc -b which re-compiled the backend WITH NO TSCONFIG.json. The reasoning for this is no path was specified for the tsc -b command.
This lead to npm run build attempting to build with es5/default target version. Causing new errors on building to appear as they are more modern features the old version does not support, and were not displayed in the editor as ESLint/Typescript was using the updated version for linting/error detection.

Also this change removed the double compiling of the electron code, first with what we actually wanted and then compiling it again with tsc -b.