Now sends progress from ipc to EditBackground.
Now sends the progress update from the save to the EditBackground component. For now it just logs [10], [11], [12] ... etc. (as it 10%, 11%, 12%). Also it seemed like it was quite spammy with the webcontents.send, so it has been limited to only send it once per % increase. (saw > dozens of updates with the same %).