Improved getBackgroundType.
Based the new getBackgroundType off the fetchFromSettings() in Background.tsx.
It now gets the id, gets the filetype for the background, if shortcut then resolves it, then returns the true filetype.

This fixes the old implemenation which failed to work on shortcut files.
Originally rewrote this in DesktopGrid.tsx, but re-wrote it in ipc before commiting to avoid the await ipc call time.