Maintaining states for hiding/showing video controls.
Now maintains the states (playing, progress, volume) on hiding and then re-showing the video controls.

For instance before, if you paused it and muted it, then hid the controls, and showed them again it would always show as playing (despite not playing), unmuted (despite being muted), and the video seek progress at 0/video length despite the current timer next to it not being 0:00.

This now preserves the states, and re-shows it correctly on re-showing video controls.

Also audio level now persists across backgrounds instead of resetting when swapping backgrounds.

Not 100% set on the setup for audio in general in this video controls. I am thinking of just making it a "temporary override" such that when Audio control is added to the background-details-panel but you are going through backgrounds seeking through you can have a temporary "master" volume. Instead of having to set the volume on every background you select.