Pause Video on minimizing/hiding, play on restore/showing.
Now pauses the video playback when minimized, keybind minimized, or X closed (closed to tray).

And resumes the video when un-minimizing, "show" in tray menu, or keybind restoring.

known issue: this means that even if the video was paused on purpose by the user it will unpause the video on restore/un-minimizing. Will implement a way to maintain this state if the user actually manually paused it.