Removed lastBounds.
I think this was an old remnant of early in the project, not entirely sure why this was implemented. Maybe something to do with the keybind restoring in an attempt to make it show on top if not focused first, then if on top minimize. Not 100% sure, but it's existence creates bugs, it was either always update lastBounds even for "minimize" or just remove it, for now it has been removed for simplicity. I just hope I don't come across the reason why it was implemented later... (did a good amound of testing and found no problems with showing/hiding, minimizing/un-minimizing).