Added local.volume (localVolume) bg.json values.
Defaults to 0.5 on save.

Not fully implemented, a lot of problems found trying to implement it so just updating the types and setting for now. Will take a lot of work to make it work as expected.
For now creating a new background will add the "volume": 0.5 under "local" in bg.json.