showVideoControls -> showVIdeoControlsRef
Swapped the type for showVideoControls to a useRef instead of useState. Spent way too long trying to figure out what was wrong with it not realizing it was resetting due to being reset via re-render on background-preview. It now is a useRef which maintains its state.

Still need an actual way to get the volume setting from the bg.json file, so it is still not fully implemented.

For now: starting the program mutes it, showing the video controls you can then un-mute it, or just open BackgroundSelect and it will set the volume to 1 (any preview sets the volume to 1). This is just for testing and making sure it was keeping the states.