Background audio now loads from bg.json (or 0.5)
Now loads the audio from the bg.json "local" { "volume": x.x;} setting.

If no local.volume exists it just defaults to 0.5.

Will add a way to actually change this default volume soon.
As of now you can test by creating a new background, then changing the volume setting in the bg.json manually. (0.00 to 1.00).

This works as long as you are not showing video controls. It is expected that if the user has video controls open they are using it like a master audio track. For instance if you want to flip through a lot of backgrounds without hearing all the audio you can just open video controls and mute it. Then you can go through all the backgrounds you want, and not have to hear all the audio as you flip through them (could be quite annoying).