videoControls position now retains position on swapping backgrounds.
Swapped from a useState which reloaded every mount to use a useRef stored in the parent to remember where it was on previewing/swapping backgrounds.

Note: this does accidentally mean that unchecking showVideoControls and re-checking it will re-draw the video controls in the same location instead of resetting it back to the bottom (this was a failsafe to re-show the video controls in case they got set off screen etc.). Will need to update this to only reset the position when toggling show video controls to true.