Fix showControls sometimes taking a few seconds to show.
Thought useRef was originally required, turns out useState is fine for videoControls. Not sure why I was having problems with it originally. Through basic testing it was working correctly keeping the video controls shown, through swapping backgrounds even to non-video backgrounds and then back to video backgrounds etc. Did not seem to be getting reset like it was. Think I fixed a problem which was causing it to reset over the last two days.