Will now jump for removedIds as well.
Will now detect removedIds and jump as well if bgs are removed.

Known issue: Jumping can actually jump to the original location. It seems like with the timing of index finishing can line up with the jump thus it will jump to the original location....  will need to work with it more.