VolumeSlider now saves volume to bg.json.
Volume slider in BackgroundSelect now saves volume to bg.json file. This means that when the backgroundselect is closed, it will update the volume. (also works when swapping backgrounds and back). Needs more work to make it consistent and live updating.

Also noticed some bugs with the volume appearance in the BackgroundSelect window when opening it, or swapping and swapping back, so evidently it is not fully up to date.