Fixed volume setting in bg.json NOT setting volume level.
Fixed the logic for setting volume.

Originally it would read the value for volume from bg.json, then if it was 0 it would set it mute, OTHERWISE it would set it to 0.5.
Didn't do much testing and was more focused with getting the slider to work so didn't realize it just always had everything defaulting to 0.5.

It now tries to set the volume to the setting, then defaults to 0.5 if volume setting is not a number or non-existent.