Volume slider now previews new volume.
Volume slider now updates the volume of the background.

Note: Volume slider is not 100% set in stone, it has a 300ms debounce which requires the user to stop changing the volume for 300ms before it will save the volume (and update the volume). Not 100% sure how I will fix that to be a bit faster. Reasoning for this is writing to the file too often causes a rare bug which duplicates "}" which breaks the bg.json entirely, reverting it to a blank state. Probably just need to backload the actual saving to file and instead use previews. Will need to do some tinkering around with it.