Better background Volume live previewing. BackgroundPreviewUpdate
Big update:

backgroundPreviewUpdate now takes a "BackgroundPreviewUpdate" type. (see types.d.ts).
This will essentially be the background logic also stored in bg.json (for now its just "id" and "volume").

Updated calls to use new BackgroundPreviewUpdate type.

Now when background.tsx receives the update-background-preview message, it will update the background (id) if passed with the id. If not then it will set the volume if passed with it.

This strategy avoids the problems with updating the bg.json directly every time the user changes the volume slider (too many too fast corrupted the bg.json). And now allows for live previewing the new volume.