Improved volume updating.
Fixed a problem where even though it had a timeout of 300ms, it would queue multiple inputs if done quick enough to save at the same time, leading to problems with writing to the bg.json file corrupting the bg.json. It now attempts to save the bg.json every 400ms. So long as changes have been made.