Improved logging and prevent resetting corrupted jsons.
Cut out a few spamming log statements for scrolling back and forth for volume.

Also when a bg.json fails to read (corrupted) it no longer defaults everything, it instead just returns. This is to save the data rather than just reset without telling the user (as the reset might not even be apparent due to previewing the background etc.).