Fixed volume being saved not being up to date.
Fixed a problem where it would queue up the volume to save at the beginning of that 400ms, but by then the value had changed it would not reflect the new volume, but rather the volume it was queued with.

This now has a up to date reference when saving volume.