Fixed swapping backgrounds setting volume (showVideoControls problems).
Fixed a problem with the useState for showVideoControls.

Basically the problem is that showVideoControls is a useState which means that it resets every time the component is redrawn (like when update-background-preview messages are received which swap the background).

The problem with this is that we need showVideoControls for rendering to instantly cause it to be drawn. If we do not have it as a useState it will not immediately show. But we also need to persist the boolean of whether it is showing such that we can handle stuff like volume changes to use the VideoConrtrols volume vs the default saved volume.

Long story short we need both :(