Video resumes when VideoControls are hidden.
Since we revert the volume it also makes sense to resume playing the video on hiding videoControls.

Also this fixes an unintended bug. The bug was if you paused the video, then hid the videoControls, and attempted to re-open the video controls, it would render the videoControls in the -100, -100 default position, and it would not use the useEffect to move it until an update (like a new frame from an unpaused video to move). Thus it would appear off the screen, forever.