BackgroundSelect now utilizes showVideoControls.
BackgroundSelect now places a red box around the volume if showVideoControls is true. (Not 100% finished just an example use). The idea for this is that if it is showing video controls the volume slider does not do anything as the video controls volume slider is seen like an override.

Still working on the ideology behind what I want it to do, whether or not the volume slider in BackgroundSelect affects the videoControls.