Small Disabled button for volume when VideoControls is showing.
Added a small buttton "Overriden" when Video controls are shown. And locks the volume bar. This is to prevent users from accidentally changing their volume thinking it's not working because the video controls actually overrides it.

Not 100% set in stone on what happens when clicking the button. Ideally it would allow updating the volume, but would need some special case logic for it as by default it currently ignores the background-preview-updates for volume if it has VideoControls showing.