Volume Bar is now red when showVolumeControls.
Decided against a enable/disable button and instead chose to just make the bar red depending on if Video Controls are currently showing.

Ideally then when the user moves it, it sends an override volume update which adjusts the volume of the background (same as updating the volumeSlider in VideoControls.