Added bgJson to "getInfoFromID". And renamed "backgroundFolder" to "folderPath".
Added bgJson to getInfoFromID().

Also realized that the backgroundFolder was the same as the idea behind the IDInfo "folderPath". As such I have chosen to just stick with "folderPath" and updated the references to backgroundFolder.

I felt folderPath is a better descriptor for what it is, as I had thought maybe backgroundFolder referred to the actual folder the background image/video is in (since .lnks are possible). But no it referred to the /backgrounds/ (if saved in default) or the external directories