EditBackground now shows red slider when VideoControls are showing.
The red slider now mirrors how it works in BackgroundSelect (showing red when videoControls are overriding the value).

Still not 100% set on just showing it as red and calling it a day, also want to better include showVideoControls in the menus like BackgroundSelect and EditBackground, where they are very convenient but would require closing out these windows to show normally.