Improper background file paths now remove tags.
Thought process behind this is if the user accidentally breaks the background path its best to just remove the tags as to not allow it to save with a video/image or a resolution. Then if possible we could maybe institute a way to look for backgrounds which do not have an image, or resolution tag under the idea that they are "broken" i.e. not a working background.