Gif type backgrounds now tag automated (and image).
gifs are not ideal compared to videos as they are extremely lossy and poor colors. However it is one of the obvious triggers to get an "animated" background.

To be honest, not super happy about supporting gif backgrounds in general because of the above problems with gifs. Contemplated just removing "Animated" tag, but its probably best to just support it and allow a tool for filtering it. After all .gifs can be renamed to .pngs and still work as gifs, just without the .gif extension. (apngs).