Updated catching uncaught errors.
Thought I had a setup to catch all main process errors, turns out ipcHandlers handle uncaught errors a bit differently.

Now correctly logs any errors that occur in ipc handlers.

For instance as of right now it is possible to create a new background with a name like "32:9" Where the ":" is an illegal character, thus when trying to save to the "32:9" folder in (saveToBackgroundIDFile ipc) it has an uncaught error.

This is now logged to the log file. Previously it just stopped the user from saving and only logged:
[ipcHandlers.ts] SaveToBackgroundIDFile called with: {"id":"32:9","sourcePath": ... }
(Did not give proper information on why it failed to save and close the EditBackground window.)