Show icons -> Hide icons.
Changed from a default (true) to default (false). Important to keep settings like this defaulted in either True or False, so a user can easily tell when a setting is not-normal. For example the Display in BackgroundSelect, it will have the following:
"Hide Icons"
"Hide Icon Names"
"Show Video Controls"

The three of which will default to unchecked (false). This will be much more intuitive that these are settings which can be turned on, and gives the user the idea that the default state is off.