Moving toward unified RendererStates.
Added hideIcons and hideIconNames to rendererStates.

Instead of having specific ipc calls like "showVideoControls" we can instead just update the renderer-state, then the windows that use that rendererState can react to the renderer updated message to update their state. This helps keep the important renderer updates on the same page between multiple windows.

Steps to accomplishing this:
Components should load the info they need on mount like such:
  useEffect(() => {
    const updateRendererStates = async () => {
      const rendererStates = await window.electron.getRendererStates();
      setHideIcons(rendererStates.hideIcons || false);
      setHideIconNames(rendererStates.hideIconNames || false);
      setShowVideoControls(rendererStates.showVideoControls || false);
    };
    updateRendererStates();
  }, []);

To actually receive and update say for the desktop which only cares about the "showVideoControls" message it can look like this:

  useEffect(() => {
    const showControls = (...args: unknown[]) => {
      logger.info("renderer-state-updated event received:", args[1]);
      const state = args[1] as Partial<RendererStates>;
      if ("showVideoControls" in state) {
        if (!state.showVideoControls) {
          videoControlsPositionRef.current = { x: -100, y: -100 };
          setVolumeFromDefault();
          const video = videoRef.current;
          video.play();
        }
        setShowVideoControls(!!state.showVideoControls);
      }
    };
    window.electron.on("renderer-state-updated", showControls);
    return () => {
      window.electron.off("renderer-state-updated", showControls);
    };
  }, []);

Then changing a rendererState is as simple as:
    await window.electron.setRendererStates({
      showVideoControls: !showVideoControls,
    });

Effective 3 steps need to be maintained:
1. Components load their needed states on launch.
2. Components listen for renderer-state-updated and update their states if any they track are updated.
3. Enacting a change is as simple as just calling setRendererStates({state: value}).

This allows for up to date rendererStates to be shared across components.

Note: not fully implemented, mostly just worked to get a proof of concept with showVideoControls syncing.