Background rendererStates cleanup.
Organization for Background.tsx to match the redererStates from other components, and added function to load (don't believe it needed it as its always defaulted on start and background is one of the first things to load, but just in case something re-renders).

Also renamed onMount fetching rendererStates to "fetchRendererStates" instead of "updateRendererStates" which might get confused with the renderer-state-updated .send calls.