Prep work for [row,col] Icons -> icon ids (for images).
Thought is to move away for strictly limiting and referring to icons and how their images are stored as based entirely on [row,col] and will instead move toward a solution where the icons have ids and draw their images from the ids.

Reasoning for this is it is required for any future work if say I wanted to implement background profiles, where the user could create a profile for a certain background, and that profile controls where icons are, what icons are shown etc. For instance say you have some backgrounds that nothing really is going on in the middle so you put some icons there, then you also have other backgrounds you like using, but icons in the center on these are clunky as theres a lot going on there. So ideally you can create profiles which these backgrounds can save with where icons are in different positions, some icons are hidden etc.

And so if doing this we can't just store icons images entirely in a strict [row,col] instead we should create an ID system where it can easily retreive its icons.

As of now, added an "id" field to new icons (it just saves as "" for now, so avoid using it for now). Also added ensureUniqueIconID which just returns the string passed to it.