EditIcon now adds ids to icons on save, and ensures unique icon ids.
Compares the oldId to the name, if the name does not match the baseName of the oldId (name has changed) it generates a new unique id. If the old id is "" (new icon) and the name has not changed, it generates a new unique id based on "unknownIcon".

Lastly if the name has not changed it leaves the id.

If it fails to read from the desktopIcons.json, it displays an error and aborts save (to avoid accidentally giving it a non-unique name which matters when we make this rename folders for example).