Prevent icon names including improper folder characters.
Removed support for specific characters that cause errors when saving as folder names. Example: /, <, >, ", etc.

There could be a way to allow this in names but it is simply best to just not allow it as the use cases for it are very rare and could complicate the process and lead to bugs.