Fixed warning on local icon image path for getting file type.
Fixed a case where it was getting the file path of a locally saved icon, thus throwing a warning because getFileType requires a full path.
This getFileType is only relevant if we are actually saving a full pathed image to local. So it has been reworked to only request file type if it determines it to be a full drive path.

Also cleared up some notes to be more clear on some functions.
Improved logging statements to help commonly logged "File does not exist" to provide more information to accurately determine which function failed in ipcHandler.