saveIconImage now saves to the /data/"icon_id" folder.
Notice: From this point until the updates with icon are done, all icon images are marked as volatile and can/will be broken over the next few days.

I will not be coding a migration, as nobody is currently using this program (that i know of).

Now saveIconImage saves to the id folder instead of the "[row,col]" folder.