SelectIconWindow, temp folder, renameDataFolder (merging).
Icon images are still heavily volatile, and are/can be broken very easily.
Recommended to just rebuild your desktop from scratch in a few days when this whole process is completed.

A massive step towards the new icon situation.

Major changes:
1. A new icon is now given a unique id starting with "temp" i.e. if it exists it becomes temp_1 etc.

This Id allows it to create a temp folder for generate Icon. Which it then stores the generated icons. If it is eventually given a name it will rename the folder to the new id on save, before saving the selected icon to that folder (if it was not the generated one).

2. Generate icon now saves into the /data/id folder.

3. SafeImages now properly display images in SelectIconWindow. (Before would only display images which were full file path or local).

Still a bunch of work that needs to be done:
Deleting the /data/id folder (currently doesn't).
Proper previewing for some types, (SelectIconWindow when you select a generatedIcon it fails to display (localPath)).

And a whole lot more....